feat: 테스트룸 관련 예외처리 구현

+++ src/main/java/algo_arena/room/service/RoomIOService.java
@@ -1,16 +1,19 @@
 package algo_arena.room.service;
 
+import static algo_arena.common.exception.enums.ErrorType.*;
 import static algo_arena.room.enums.RoomEvent.*;
 
 import algo_arena.member.entity.Member;
 import algo_arena.member.service.MemberService;
 import algo_arena.room.entity.Room;
+import algo_arena.room.exception.RoomException;
 import algo_arena.room.repository.RoomRedisRepository;
 import algo_arena.room.repository.RoomRepository;
 import algo_arena.room.service.result.RoomEventResult;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.StringUtils;
 
 @Service
 @RequiredArgsConstructor
@@ -23,11 +26,8 @@ public class RoomIOService {
     @Transactional
     public RoomEventResult enterRoom(String roomId, String memberName) {
         Room room = getRoomFromDB(roomId);
-        if (room.isMember(memberName)) {
-            throw new RuntimeException();
-        }
         if (room.isFull()) {
-            throw new RuntimeException();
+            throw new RoomException(ROOM_FULL);
         }
         Member member = memberService.findMemberByName(memberName);
         room.enter(member);
@@ -38,10 +38,11 @@ public class RoomIOService {
     public RoomEventResult exitRoom(String roomId, String memberName) {
         Room room = getRoomFromDB(roomId);
         if (!room.isMember(memberName)) {
-            throw new RuntimeException();
+            throw new RoomException(NOT_IN_ROOM);
         }
+
         if (room.isHost(memberName) && !room.existMembers()) {
-            deleteRoom(roomId);
+            deleteRoom(room.getId());
             return RoomEventResult.from(DELETE);
         }
         if (room.isHost(memberName)) {
@@ -52,17 +53,13 @@ public class RoomIOService {
         return RoomEventResult.from(EXIT, room);
     }
 
-    public void checkAlreadyEntrance(String memberName) {
-        boolean isAlreadyEntered = roomRepository.findAll().stream()
-            .anyMatch(room -> room.isMember(memberName));
-        if (isAlreadyEntered) {
-            throw new RuntimeException();
-        }
-    }
-
     private Room getRoomFromDB(String id) {
+        if (!StringUtils.hasText(id)) {
+            throw new RoomException(NULL_VALUE);
+        }
         roomRedisRepository.deleteById(id);
-        return roomRepository.findById(id).orElseThrow();
+        return roomRepository.findById(id)
+            .orElseThrow(() -> new RoomException(ROOM_NOT_FOUND));
     }
 
     private void memberExitRoom(Room room, String memberName) {
@@ -76,7 +73,7 @@ public class RoomIOService {
     }
 
     private void deleteRoom(String roomId) {
-        roomRepository.deleteById(roomId);
         roomRedisRepository.deleteById(roomId);
+        roomRepository.deleteById(roomId);
     }
 }
